+++ src/main/java/io/github/artsok/internal/ParameterizedTestInvocationContext.java
@@ -8,7 +8,7 @@
  * http://www.eclipse.org/legal/epl-v20.html
  */
 
-package io.github.artsok.params;
+package io.github.artsok.internal;
 
 import org.junit.jupiter.api.extension.Extension;
 import org.junit.jupiter.api.extension.TestTemplateInvocationContext;
@@ -20,7 +20,7 @@ import static java.util.Collections.singletonList;
 /**
  * @since 5.0 - COPY PAST FROM ORIGINAL JUNIT 5 WITH SEVERAL CORRECTIONS
  */
-class ParameterizedTestInvocationContext implements TestTemplateInvocationContext {
+public class ParameterizedTestInvocationContext implements TestTemplateInvocationContext {
 
     private final int currentRepetition;
     private final int totalRepetitions;
@@ -29,7 +29,7 @@ class ParameterizedTestInvocationContext implements TestTemplateInvocationContex
     private final ParameterizedRepeatedMethodContext methodContext;
     private final Object[] arguments;
 
-    ParameterizedTestInvocationContext(int currentRepetition, int totalRepetitions, ParameterizedRepeatedIfExceptionsTestNameFormatter formatter,
+    public ParameterizedTestInvocationContext(int currentRepetition, int totalRepetitions, ParameterizedRepeatedIfExceptionsTestNameFormatter formatter,
                                        ParameterizedRepeatedMethodContext methodContext, Object[] arguments) {
         this.currentRepetition = currentRepetition;
         this.totalRepetitions = totalRepetitions;
